Use `futures_timer` crate and fix unstable tests #1509 (#1511)
* Use `futures-timer` instead of custom Timer implementation

* Remove `TimedMutexGuard` as it's not used anymore

* Run `cargo test` with `--nocapture` flag on MacOS CI

* Increase timeouts at the `abortable_system` tests

* Play with the timeouts

* `Timer::sleep_ms` should take ms as u64

+++ mm2src/coins/lightning/ln_events.rs
@@ -509,7 +509,7 @@ impl LightningEventHandler {
 
     fn handle_pending_htlcs_forwards(&self, time_forwardable: Duration) {
         info!("Handling PendingHTLCsForwardable event!");
-        let min_wait_time = time_forwardable.as_millis() as u32;
+        let min_wait_time = time_forwardable.as_millis() as u64;
         let channel_manager = self.channel_manager.clone();
         self.platform.spawner().spawn(async move {
             let millis_to_sleep = rand::thread_rng().gen_range(min_wait_time, min_wait_time * 5);
